Fixed a flow issue
