+++ src/components/Nav/DashboardNav.jsx
@@ -1,17 +1,25 @@
 import React, { useState } from "react";
 import {
   Modal,
+  ModalHeader,
+  ModalBody,
+  Button,
+  ModalFooter,
   UncontrolledDropdown,
   DropdownToggle,
   DropdownMenu,
   DropdownItem,
 } from "reactstrap";
 import EventInvitation from "../EventInvitation";
+import EditUser from '../forms/EditUser'
 import { Link } from "react-router-dom";
 
 export default function DashboardNav() {
   const [modal, setModal] = useState(false);
   const toggleModal = () => setModal(!modal);
+  const [userModal, setuserModal] = useState(false);
+  const userToggle = () => setuserModal(!userModal);
+
   return (
     <div>
       <UncontrolledDropdown nav inNavbar>
@@ -19,8 +27,8 @@ export default function DashboardNav() {
           AfterLoginMenu
         </DropdownToggle>
         <DropdownMenu right>
-          <DropdownItem>
-            <Link to="/profile">Profile</Link>
+          <DropdownItem onClick = {userToggle}>
+            Update Profile
           </DropdownItem>
           {/* map over upcoming events with dropdown and dynamic link */}
           <DropdownItem>
@@ -36,6 +44,20 @@ export default function DashboardNav() {
 
       <Modal isOpen={modal} toggle={toggleModal}>
         <EventInvitation />
+      </Modal>
+
+      <Modal isOpen={userModal} toggle={userToggle}>
+        <ModalHeader toggle={userToggle}>
+          <h2>Update Information</h2>
+        </ModalHeader>
+        <ModalBody>
+          <EditUser />
+        </ModalBody>
+        <ModalFooter>
+          <Button className="bg-cancel" onClick={userToggle}>
+            Cancel
+          </Button>
+        </ModalFooter>
       </Modal>
     </div>
   );
